refactor(test): group networking coverage
