+++ test/networking/compatible-endpoint-context-probe.test.ts
@@ -12,13 +12,13 @@ import { afterEach, describe, expect, it } from "vitest";
 import {
   applyCompatibleEndpointContextWindow,
   fetchCompatibleEndpointModels,
-} from "../src/lib/inference/compatible-endpoint-context";
+} from "../../src/lib/inference/compatible-endpoint-context";
 import {
   type FakeOpenAiCompatibleServer,
   startFakeOpenAiCompatibleServer,
-} from "./e2e/fixtures/fake-openai-compatible";
-import { startTestProgress, type TestProgress } from "./e2e/fixtures/progress.ts";
-import { testTimeout } from "./helpers/timeouts";
+} from ".././e2e/fixtures/fake-openai-compatible";
+import { startTestProgress, type TestProgress } from ".././e2e/fixtures/progress.ts";
+import { testTimeout } from ".././helpers/timeouts";
 
 const MODEL = "nvidia/NVIDIA-Nemotron-3-Super-120B-A12B-NVFP4";
 let progress: TestProgress | null = null;
