+++ test/networking/http-proxy-fix-sync.test.ts
@@ -7,7 +7,7 @@ import os from "node:os";
 import path from "node:path";
 import { describe, expect, it } from "vitest";
 
-const ROOT = path.join(import.meta.dirname, "..");
+const ROOT = path.join(import.meta.dirname, "../..");
 const CANONICAL_FIX = path.join(ROOT, "nemoclaw-blueprint", "scripts", "http-proxy-fix.js");
 const START_SCRIPT = path.join(ROOT, "scripts", "nemoclaw-start.sh");
